Desktop/Image_Tagger_WebApp/requirements.txt
Flask
google-generativeai
Pillow
pillow-heif
pandas